Moves SearchTextField to own file
Reduces the size of TokenListViewController
SearchTextField functionality is mostly self-contained
