+++ Authenticator/Source/TokenListViewController.swift
@@ -308,95 +308,3 @@ extension TokenListViewController: UITextFieldDelegate {
     }
 
 }
-
-// MARK: Custom search field for navigation bar
-
-class SearchField: UIView {
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        setupTextField()
-    }
-
-    required init?(coder aDecoder: NSCoder) {
-        super.init(coder: aDecoder)
-        setupTextField()
-    }
-
-    var delegate: UITextFieldDelegate? {
-        get { return textField.delegate }
-        set { textField.delegate = newValue }
-    }
-
-    var text: String? {
-        get { return textField.text }
-    }
-
-    let ring = OTPProgressRing(
-        frame: CGRect(origin: .zero, size: CGSize(width: 22, height: 22))
-    )
-
-    let textField: UITextField = SearchTextField()
-
-    private func setupTextField() {
-        ring.tintColor = UIColor.otpLightColor
-        let placeHolderAttributes = [
-            NSForegroundColorAttributeName: UIColor.otpLightColor,
-            NSFontAttributeName: UIFont(name: "HelveticaNeue-Light", size: 16)!
-        ]
-        textField.attributedPlaceholder = NSAttributedString(
-            string: "Authenticator",
-            attributes: placeHolderAttributes
-        )
-        textField.textColor = UIColor.otpLightColor
-        textField.backgroundColor = UIColor.otpLightColor.colorWithAlphaComponent(0.2)
-        textField.leftView = ring
-        textField.leftViewMode = .Always
-        textField.borderStyle = .RoundedRect
-        textField.clearButtonMode = .Always
-        textField.autocorrectionType = .No
-        textField.autocapitalizationType = .None
-        textField.keyboardAppearance = .Dark
-        textField.sizeToFit()
-        addSubview(textField)
-    }
-
-    override func intrinsicContentSize() -> CGSize {
-        return ring.frame.union(textField.frame).size
-    }
-
-    override func sizeThatFits(size: CGSize) -> CGSize {
-        let fieldSize = textField.frame.size
-        let fits = CGSize(width: max(size.width, fieldSize.width), height: fieldSize.height)
-        return fits
-    }
-
-    override func layoutSubviews() {
-        super.layoutSubviews()
-
-        // center the text field
-        var textFieldFrame = textField.frame
-        textFieldFrame.size.width = bounds.size.width
-        textField.frame = textFieldFrame
-        textField.center = CGPoint(x: bounds.size.width * 0.5, y: bounds.size.height * 0.5)
-    }
-
-}
-
-class SearchTextField: UITextField {
-    override func leftViewRectForBounds(bounds: CGRect) -> CGRect {
-        return super.leftViewRectForBounds(bounds).offsetBy(dx: 6, dy: 0)
-    }
-
-    override func placeholderRectForBounds(bounds: CGRect) -> CGRect {
-        if self.isFirstResponder() {
-            return .zero
-        }
-        var rect = super.placeholderRectForBounds(bounds)
-        if let size = attributedPlaceholder?.size() {
-            rect.size.width = size.width
-            rect.origin.x = ( bounds.size.width - rect.size.width ) * 0.5
-        }
-        return rect
-    }
-}
